@@ -12,7 +12,7 @@ class NewsController < ApplicationController
   # GET /news.json
   ##
   def index
-    @news = News.paginate(:page => params[:page], :per_page => 5).public
+    @news = News.paginate(:page => params[:page], :per_page => 5).for_public
     @title = "News"
     respond_to do |format|
       format.html # index.html.erb
