@@ -5,7 +5,7 @@ class StaticController < ApplicationController
     if current_user
       @news = News.current.limit(2).all
     else
-      @news = News.public.limit(2).all
+      @news = News.for_public.limit(2).all
     end
     @shows = Show.current_slideshow.limit(5).all
     @last_show = Show.last_show
