@@ -23,10 +23,10 @@ class Admin::Staffing < ActiveRecord::Base
   after_save     :update_reminder
   before_destroy :reminder_cleanup
 
-  default_scope order("start_time ASC")
+  default_scope -> { order("start_time ASC") }
 
-  scope :future, where(['end_time > ?', DateTime.now])
-  scope :past, where(['end_time < ?', DateTime.now])
+  scope :future, -> { where(['end_time > ?', DateTime.now]) }
+  scope :past, -> { where(['end_time < ?', DateTime.now]) }
 
   has_many :staffing_jobs, :as => :staffable, :class_name => "Admin::StaffingJob", :dependent => :destroy
   has_many :users, :through => :staffing_jobs
