@@ -20,7 +20,7 @@
 #++
 ##
 class TeamMember < ActiveRecord::Base
-  default_scope order("display_order ASC")
+  default_scope -> { order("display_order ASC") }
 
   belongs_to :teamwork, :polymorphic => true
   belongs_to :user
